+++ frontend/src/components/common/HeaderCategoriesComponent.vue
@@ -79,16 +79,15 @@
                 v-for="(bottoms, index) in bottomCategoriesLists"
                 :key="index"
               >
-                <div
+                <router-link
                   v-for="bottom in bottoms"
                   :key="bottom.idx"
-                  @click="moveToProductDetail(bottom.idx)"
-                >
+                  :to="{ name: 'categoryProductList', params: { categoryIdx: bottom.idx },}">
                   <span
                     class="w-full px-[12px] py-[8px] flex items-center gray-333--text body1-regular-small cursor-pointer shrink-0 hover:underline"
                     >{{ bottom.categoryName }}</span
                   >
-                </div>
+                </router-link>
               </div>
             </div>
           </div>
@@ -197,10 +196,11 @@ export default defineComponent({
     // console.log("부모카테고리", categoryStore.parentCategory);
 
     //소분류 클릭시 이동하기
-    const moveToProductDetail = async (categoryIdx) => {
-      router.push(`/category/${categoryIdx}`);
-      closeCategory();
-    };
+    // const moveToProductDetail = async (categoryIdx) => {
+    //   router.push(`/category/${categoryIdx}`);
+    //   closeCategory();
+    // };
+    
 
     const closeCategory = () => {
       emit("closeCategory");
