+++ frontend/src/components/product/ProductList4LayoutComponent.vue
@@ -7,7 +7,6 @@
                 <ProductComponent :product="product"/>
             </router-link>
         </div>
-         <!-- <ObserverComponent @show="infiniteHandler"></ObserverComponent> -->
     </div>
 </template>
 
@@ -15,12 +14,10 @@
 import { mapStores } from "pinia";
 import { useProductStore } from "@/stores/useProductStore";
 import ProductComponent from './ProductComponent.vue'
-// import ObserverComponent from './ObserverComponent.vue';
 
 export default {
     components:{
         ProductComponent,
-        // ObserverComponent
     },
     data(){
         return{
@@ -38,29 +35,8 @@ export default {
         ...mapStores(useProductStore)
     },
     created(){
-        // this.productStore.productList = [];
     },
     methods:{
-        async searchByCategory(){
-            await this.productStore.searchByCategory(1,this.page, 12);
-        },
-        async infiniteHandler(){
-            if (this.loading) return; 
-            this.loading = true; 
-
-            if (this.page !== 0) {
-                await new Promise((resolve) => setTimeout(resolve, 150));
-            }
-
-            try {
-                await this.productStore.searchByCategory(1,this.page, 12);
-                this.page++;
-            } catch (error) {
-                console.error("Failed to fetch data:", error);
-            } finally {
-                this.loading = false; 
-            }
-        }
     }
 }
 </script>
