+++ frontend/src/pages/product/CategoryProductListPage.vue
@@ -3240,6 +3240,7 @@
             </div>
             <!-- 상품 리스트 -->
             <ProductListComponent :productList="productStore.productList"/>
+            <ObserverComponent @show="infiniteHandler"></ObserverComponent>
           </div>
         </div>
       </div>
@@ -3326,11 +3327,13 @@ import { useProductStore } from "@/stores/useProductStore";
 import { useRoute } from "vue-router";
 import SideBarComponent from "@/components/product/SideBarComponent.vue";
 import ProductListComponent from "@/components/product/ProductList4LayoutComponent.vue";
+import ObserverComponent from '@/components/product/ObserverComponent.vue';
 
 export default {
   components: {
     SideBarComponent,
     ProductListComponent,
+    ObserverComponent
   },
   computed:{
         ...mapStores(useProductStore)
@@ -3338,6 +3341,8 @@ export default {
   data() {
     return {
       categoryIdx: null, // 초기값 설정
+      page:0,
+      loading: false //로딩 관리
     };
   },
   created() {
@@ -3348,21 +3353,38 @@ export default {
     // this.getProductListByCateory(this.categoryIdx, 0, 12);
   },
   methods:{
-      async getProductListByCateory(idx, page, size){
-        await this.productStore.searchByCategory(idx, page, size);
-      }
+      // async getProductListByCateory(idx, page, size){
+      //   await this.productStore.searchByCategory(idx, page, size);
+      // },
+      async infiniteHandler(){
+            if (this.loading) return; 
+            this.loading = true; 
+
+            if (this.page !== 0) {
+                await new Promise((resolve) => setTimeout(resolve, 150));
+            }
+
+            try {
+                await this.productStore.searchByCategory(this.categoryIdx,this.page, 12);
+                this.page++;
+            } catch (error) {
+                console.error("Failed to fetch data:", error);
+            } finally {
+                this.loading = false; 
+            }
+        }
   },
-  watch: {
-    "$route.params.categoryIdx": {
-      immediate: true,
-      handler(newCategoryIdx) {
-        this.categoryIdx = newCategoryIdx; // 카테고리 ID 갱신
-        console.log("변경된 카테고리 ID:", this.categoryIdx); // 확인용 로그
-        this.getProductListByCateory(this.categoryIdx, 0, 12);
-        console.log(this.productStore.productList);
-      },
-    },
-  },
+  // watch: {
+  //   "$route.params.categoryIdx": {
+  //     immediate: true,
+  //     handler(newCategoryIdx) {
+  //       this.categoryIdx = newCategoryIdx; // 카테고리 ID 갱신
+  //       console.log("변경된 카테고리 ID:", this.categoryIdx); // 확인용 로그
+  //       // this.getProductListByCateory(this.categoryIdx, 0, 12);
+  //       // console.log(this.productStore.productList);
+  //     },
+  //   },
+  // },
 };
 </script>
